requirements.txt
pymodbus
psycopg2-binary